+++ state-chain/traits/src/mocks/epoch_info.rs
@@ -1,5 +1,5 @@
 pub type Mock = MockEpochInfo;
-crate::impl_mock_epoch_info!(u64, u128, u32, u16);
+crate::impl_mock_epoch_info!(u64, u128, u32, u32);
 
 #[macro_export]
 macro_rules! impl_mock_epoch_info {
